requirements.txt
ctransformers
sentence_transformers>=2.6.0
pinecone-client
langchain
flask
pypdf
transformers
torch
langchain_huggingface
python-dotenv
-e .